More WIP - now has an emailing test
Updated to create an emailer test file that is capable of
parsing and processing the templates and sending an email.

Also updated the templates to use bootstrap and to rework them to
better match what is required.

Also updated the root commands and the config packages to add more
test cases and some refactoring.

Added a couple of bad config fie examples for testing.