Revised to use the github.com/emersion/go-smtp smtp package
Revised the codebase to make use to the 3rd party
github.com/emersion/go-smtp smtp and github.com/emersion/go-message/mail
mail packages.

These replace the previous use of the fork of the
github.com/jordan-wright/email package.

The config data has been revised to support a new "Auth" section that
holds the username and password required for the smtp server.

The email that is sent contains both a HTML and text part.

The tests have been updated to reflect the changes.

At the minute TestSendMail will only send a "customer" thank you email.
The system email which contains the feedback provided by the user has
not yet been ported.