Update config.ReadConfig to use viper to find the config file
Updated config.ReadConfig so that it correctly uses viper to search for
the actual config file in a number of different paths,

config.SetConfigFile will now add the paths "." and
"/etc/emailformgateway" and config file type of "TOML". If no file name
is given then then "config" is used. i.e. a file called "config.toml" is
searched for in the "." and "/etc/emailformgateway/" paths.

If a filename is specified then config.SetConfigFile will attempt to
split the supplied filename into the directory, the extension and the
filename without extension. It will then add the path, and extension and
filename to the viper search path. So a filename of
"/etc/look/here/newconfig.json" would add "/etc/look/here" as a path,
"json" as a extension and "newconfig" as a filename.

The approach used to achieve thisis not full proof, as it is really
intended to help testing.