Change the SendEmail test to bind config keys to env vars via viper
Changed the SendEmail test so that the System.To email address is bound
to the TEST_SYSTEM_TO_EMAIL_ADDRESS and the Teampates.Dir key is bound
to the env var TEST_TEMPLATES_DIR. This allows for the test to overload
these values without putting them in the config file.

The TEST_CUSTOMER_TO_EMAIL_ADDRESS will set the To address in the form
data. The TEST_CONFIG_FILE will set the config file to read. The value
should be the filename minus any extension and path e.g. "test_config"
for "./test_config.toml".

Neither of these can be bound to config keys via viper, and must be
supplied separately

All other values will be taken from the config file as normal.