Update the gatewayHandler to use env vars for config
Updated the server's HTTP handler function server.gatewayHandler to bind
the "Address.SystemTo" and "Templates.Dir" keys to the env vars
TEST_SYSTEM_TO_EMAIL_ADDRESS and TEST_TEMPLATES_DIR respectively.

This allows both to be overloaded for testing.

In addition the env var TEST_CONFIG_FILE is used to specify the config
file to read. This is NOT taken from the config held by the server, as
the handler runs in its own go routine and viper is not go routine safe.
So it is simpler to re-read the config file in the handler.

The config file reading in the server will need to be refactored at a
later date.

A new test TestServerSendEmail has been added. This test sends to a test
server by making an HTTP POst request to the server, and passing the
JSON encoded form data. This test relies on an env var
TEST_CUSTOMER_TO_EMAIL to set the email address to send the customer
"thank you" email to. This avoids hard coding the email adrress in the
test. The test confrims that the client would recieve an HTTP 200/OK.