Fix: TestServerSendEmail so that it correctly checks the server response
Fix the test TestServerSendEmail so that it corretly reads the response
body and confirns that the response is a JSON encoded formResponse.

The returned formResponse should have the Valid field set to true and
BadFields set to nil i.e. en empty array. If these conditions are not
met th test fails.