Update TestReadConfigWithEnvVarOverload to only use env var if defined
Update TestReadConfigWithEnvVarOverload such that if the env var
TEST_PASSWORD is defined, then that value (assuming its not an empty
string) will be used as the expected value of the Auth.Password key.

Theis stops the test from failing, if the env var isn't defined, but
respects it if its defined.

In the case where the env var is not defined, then TestReadConfigWithEnvVarOverload
is equivalent to TestReadConfig.