Replace the use of strings.Title which is deprecated
strings.Title is deprecated becuase it didn't handle Unicode upper-casing
correctly. It has been replaced the golang.org/x/text/cases package and
the new cases.Title method.

Cases.Title correctly handles Unicode and is multi-lingual. cases.Title
will always uppercase the first letter of a word and lowercase the
remaining letters, regardless of case. There is a new test
TestCreateFromDataMap that demonstrates this, as well as confirming that
createFromDataMap works as expcted after the change to cases.Title.