Add the host, port and route as cmd line flags to the server executable
The hostname, port and the route/URL path the server will listen
on are supplied as command line paramaters to the server executable in addition to
the config file.

This change means that the configuration file no longer needs to be read by
the server main function. These details will never change once the
server is started.

The config file is now only read in the http handler
function i.e. server.gatewayHandler. This means that the viper config
will be read once for each request. This is an acceptable trade off as
we do not expect the gateway handler to be called frequently i.e. not
every minute or faster and it means that the viper configuration does
not need to be made go routine safe, as everything will be processed in
the handlers go routine.

This change also means that the environmental variables
TEST_CONFIG_FILE,
TEST_SYSTEM_TO_EMAIL_ADDRESS,
TEST_TEMPLATES_DIR,
TEST_CUSTOMER_TO_EMAIL

can all be moved into the server test code. This
avoids referencing these testing only env vars in the main server
executable.

This is change is a prelude to removing the "Server" block from the config file.