Read the config at server startup not in the handler
Rework things so that the config is read completely into a config.Config
struct when the server first starts up. The config of the server will
not change once the server is started i.e. there will be writes to the
config - as the server does not use file watching, or env var overrides.

This means that the config values can be safely read, via viper, in each
handler func which runs in its own go routine.

This simplifies the code by removing the need to try to read the config
file at server startup to see if it exits or not, and then a second read
in the handler function to read the data.

The root command has also been updated to return an error if the command
being run fails.