Correct the TEST_CUSTOMER_TO_EMAIL env var name in the server test
Change the env var:

TEST_CUSTOMER_TO_EMAIL => TEST_CUSTOMER_TO_EMAIL_ADDRESS

in the server_test.go so that it is consistent with the env var name
used in the emailer_test.go and with the end var name and secret name
defined in the github workflow file.

This will allow the test to pass.