Update the emailer to support remote SMTP servers that are secure by default
Update the emailer so that is now works with any remote SMTP servers that
require TLS from the start - rather than negioating a secure connetion.

This is required by some SMTP servers in the wild.

This also required correcting the FROM addres passed to the SMTP server.
This should be of the form "user@host.tld" and not the
"First Last <user@host.com>". This is because the SMTP protocol will
enclose the FROM email address in <...> as required by the protocol.