Fix: correct path to test_vars.sh on the server
Correct the path to the test_vars.sh file on the server.

test_vars.sh is copied into the emailformgateway directory on the server
so the ssh lines that source the test_vars.sh file have had
the path updated to include the emailformgateway directory.

Also removed the unnecessary sudo when deleting the emailformgateway folder
from the server, as part of the cleanup.